requirements.txt
streamlit>=1.28
pandas>=1.5
scikit-learn>=1.2
matplotlib>=3.5
seaborn>=0.12
Pillow>=9.0
numpy>=1.23
